visualize_github_star/requirements.txt
requests>=2.31.0
tqdm>=4.66.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv
